+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
 
+- Header image updated with reStructuredText text
 
 ## [1.0.4] - 2018-01-31