+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - `rest` command added to README.rst
+- `rest` command added to cli.py
 
 ### Changed