+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Header image updated with reStructuredText text
 
+### Fixed
+
+- cli fixed - see #2
+
 ## [1.0.4] - 2018-01-31
 
 ### Removed