+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+    
+## [1.0.5] - 2018-02-21
+
 ### Added
 
 - `rest` command added to README.rst
@@ -20,7 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - cli fixed - see #2
-
+    
 ## [1.0.4] - 2018-01-31
 
 ### Removed
@@ -104,7 +107,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 Initial release.
 
-[Unreleased]: https://github.com/akshaybabloo/release-exporter/compare/1.0.4...HEAD
+[Unreleased]: https://github.com/akshaybabloo/release-exporter/compare/1.0.5...HEAD
+[1.0.5]: https://github.com/akshaybabloo/release-exporter/compare/1.0.4...1.0.5
 [1.0.4]: https://github.com/akshaybabloo/release-exporter/compare/1.0.3...1.0.4
 [1.0.3]: https://github.com/akshaybabloo/release-exporter/compare/1.0.2...1.0.3
 [1.0.2]: https://github.com/akshaybabloo/release-exporter/compare/v1.0.1...1.0.2